Move remained join models to service-models folder
