--- src/ggrc-client/js/models/join-models/index.js
@@ -1,7 +1,0 @@
-/*
-  Copyright (C) 2018 Google Inc.
-  Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
-*/
-
-export {default as TaskGroupObject} from './task-group-object';
-export {default as UserRole} from './user-role';
