+++ src/ggrc-client/js/models/service-models/task-group-object.js
