+++ src/ggrc-client/js/tests/permission_spec.js
@@ -6,7 +6,7 @@
 import Permission from '../permission';
 import {makeFakeInstance} from '../../js_specs/spec_helpers';
 import * as CurrentPageUtils from '../plugins/utils/current-page-utils';
-import UserRole from '../models/join-models/user-role';
+import UserRole from '../models/service-models/user-role';
 import Audit from '../models/business-models/audit';
 import {getInstance} from '../plugins/utils/models-utils';
